+++ common/src/common/Obstacle.cpp
@@ -20,8 +20,8 @@ Obstacle::Obstacle(const sf::Vector2f& position, const std::string& id, int u_id
                      utils::j3x::get<int>(RM.getObjectParams("obstacles", id), "z_index"),
                      utils::j3x::get<int>(RM.getObjectParams("obstacles", id), "frames_number"),
                      utils::j3x::get<float>(RM.getObjectParams("obstacles", id), "frame_duration")),
-        Shootable(static_cast<int>(utils::j3x::get<float>(RM.getObjectParams("obstacles", id), "endurance")) *
-                  CFG.get<int>("obstacles_endurance_factor"))
+        Shootable(utils::j3x::get<float>(RM.getObjectParams("obstacles", id), "endurance") *
+                  CFG.get<float>("obstacles_endurance_factor"))
 {
     this->changeOrigin(sf::Vector2f(utils::j3x::get<float>(RM.getObjectParams("obstacles", id), "size_x"),
                                     utils::j3x::get<float>(RM.getObjectParams("obstacles", id), "size_y")) / 2.0f +
